# update worker-dom, uncomment benchmark