--- a/package.json
+++ b/package.json
@@ -35,7 +35,7 @@
     "typescript": "4.4.3"
   },
   "dependencies": {
-    "@ampproject/worker-dom": "0.31.0"
+    "@ampproject/worker-dom": "0.32.0"
   },
   "np": {
     "branch": "main",