**summary**
Bumps `worker-dom` from 0.31 to 0.32.1, which includes performance improvements to `document.createElement`.
Results are 65x faster compilation for normal sized documents (1000 nodes) 🎉  🚀 .

before
```
Document: 1 node x 31,867 ops/sec ±4.42% (69 runs sampled)
Document: 10 nodes x 8,943 ops/sec ±9.14% (71 runs sampled)
Document: 100 nodes x 640 ops/sec ±45.72% (35 runs sampled)
Document: 1000 nodes x 8.17 ops/sec ±5.72% (25 runs sampled)
# Crashes after this
```

after
```
Document: 1 node x 98,487 ops/sec ±1.45% (92 runs sampled)
Document: 10 nodes x 38,525 ops/sec ±0.98% (86 runs sampled)
Document: 100 nodes x 5,265 ops/sec ±1.78% (92 runs sampled)
Document: 1000 nodes x 522 ops/sec ±1.29% (88 runs sampled)
Document: 10000 nodes x 27.18 ops/sec ±2.06% (48 runs sampled)
Document: 100000 nodes x 2.88 ops/sec ±3.71% (12 runs sampled)
```